tui/pyproject.toml
# Package Manager: uv (recommended)
# Install with: uv pip install -e ".[dev]"
# Or: uv sync (if using uv project management)

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crystal-tui"
version = "0.1.0"
description = "A modern TUI workstation for CRYSTAL DFT calculations"
authors = [{name = "Brian Squires", email = "brian@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["crystal", "dft", "tui", "computational-chemistry"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Physics",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "textual>=0.50.0",      # Terminal UI framework
    "rich>=13.0.0",         # Rich text and tables
    "jinja2>=3.1.0",        # Template engine (used in templates.py)
    "pyyaml>=6.0.0",        # YAML parsing (used in templates.py)
    "asyncssh>=2.14.0",     # SSH/SFTP for remote execution (Phase 2)
    "keyring>=24.0.0",      # Secure credential storage (Phase 2)
]

[project.optional-dependencies]
# Optional analysis tools (not currently imported)
# Install with: pip install crystal-tui[analysis]
analysis = [
    "CRYSTALpytools>=2023.0.0",
    "pymatgen>=2023.0.0",
    "ase>=3.22.0",
]

# AiiDA integration (Phase 3)
# Install with: pip install crystal-tui[aiida]
# Requires: PostgreSQL 12+, RabbitMQ (optional)
aiida = [
    "aiida-core>=2.7,<3.0",       # AiiDA workflow engine
    "psycopg2-binary>=2.9",       # PostgreSQL adapter
]

dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

# Full development with AiiDA
dev-aiida = [
    "crystal-tui[dev]",
    "crystal-tui[aiida]",
]

[project.scripts]
crystal-tui = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
exclude = ["tests*", "*.egg-info"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::DeprecationWarning:asyncssh.*:",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
